requirements.txt
google.generativeai==0.8.3
pillow ==11.0.0
streamlit == 1.39.0
streamlit.option.menu == 0.4.0